+++ packages/dev/CHANGELOG.md
@@ -1,3 +1,16 @@
+## @remix-pwa/dev 2.0.29-dev.2 (2023-10-07)
+
+
+### Bug Fixes
+
+* **dev:** implemented brand-new watcher for worker files/routes c067e70
+
+### Dependencies
+
+* **@remix-pwa/worker-runtime:** upgraded to 2.0.7-dev.1
+
+## @remix-pwa/dev 2.0.29-dev.1 (2023-10-04)
+
 ## @remix-pwa/dev 2.0.29 (2023-10-04)